gps: kill entire child process group on failure
Previously, if dep tried to clone a repository over ssh that contained
zero padded file modes, dep would fail to clone and then hang. You can
reproduce this failure with the following .gitconfig:

    [url "git@github.com:"]
        insteadOf = https://github.com/

    [transfer]
        fsckobjects = true

    [fetch]
        fsckobjects = true

    [receive]
        fsckobjects = true

and the following Gopkg.toml. (It is not my intention to single out
this project - I searched Github for Go projects with zero padded file
mode errors, and found this one.)

    [[constraint]]
      name = "github.com/remogatto/gospeccy"
      branch = "master"

`dep ensure` hangs because the git clone operation spins up an `ssh`
process as a child process. `cmd.Process.Kill` kills the parent `git`
operation, but not the child `ssh` operation, so it's orphaned and we
don't return properly from the function. (It's unclear to me at this
point why the ssh operation did not return when it completed.)

By sending the negative Pgrp value to the Kill() function we can kill
the entire child process group, not just the "parent child." See
https://medium.com/@felixge/killing-a-child-process-and-all-of-its-children-in-go-54079af94773
for more information.

Updates #1257.